requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.0.3
requests==2.32.3
